+++ src/hooks/useOverlap.ts
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { useWorld } from "./useWorld";
 import { CollisionEventFunction } from "../types";
 
-export const useOverlap = (id: string, handler: CollisionEventFunction) => {
+export function useOverlap<T = unknown>(id: string, handler: CollisionEventFunction<T>) {
   const { registerHandler } = useWorld();
 
   useEffect(() => {