+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added support for arbitrary bitmap scaling, via the `factor` prop on `BitmapImage`.
+- Added `entity` prop to `CollisionBox`, for passing entities to collision handlers.
 - Added `useCachedDynamicProperty` as a performance optimization.
 - Limited tick delta to a maximum of one 1/15 of a second, to avoid huge deltas on startup.
 - Removed automatic rounded of positions to whole pixels in `useElement`.